+++ inst/doc/elo.R
@@ -8,10 +8,10 @@ elo.prob(elo.A, elo.B)
 
 ## ------------------------------------------------------------------------
 wins.A <- c(1, 0)
-elo.update(elo.A, elo.B, wins.A, k = 20)
+elo.update(wins.A, elo.A, elo.B, k = 20)
 
 ## ------------------------------------------------------------------------
-elo.calc(elo.A, elo.B, wins.A, k = 20)
+elo.calc(wins.A, elo.A, elo.B, k = 20)
 
 ## ------------------------------------------------------------------------
 data(tournament)
@@ -36,13 +36,45 @@ elo.run(score(points.Home, points.Visitor) ~ team.Home + elo.Visitor,
         data = tournament, k = 20)
 
 ## ------------------------------------------------------------------------
+tournament$elo.Visitor <- 1500
+elo.run(score(points.Home, points.Visitor) ~ team.Home + elo.Visitor +
+        regress(half, 1500, 0.2),
+        data = tournament, k = 20)
+
+## ------------------------------------------------------------------------
 e <- elo.run(score(points.Home, points.Visitor) ~ team.Home + team.Visitor,
              data = tournament, k = 20)
+summary(e)
+
+## ------------------------------------------------------------------------
 head(as.matrix(e))
 
 ## ------------------------------------------------------------------------
 str(as.data.frame(e))
 
 ## ------------------------------------------------------------------------
-last(e)
+final.elos(e)
+
+## ------------------------------------------------------------------------
+results <- elo.run(score(points.Home, points.Visitor) ~ adjust(team.Home, 10) + team.Visitor,
+                   data = tournament, k = 20)
+newdat <- data.frame(
+  team.Home = "Athletic Armadillos",
+  team.Visitor = "Blundering Baboons"
+)
+predict(results, newdata = newdat)
+
+## ------------------------------------------------------------------------
+dat <- data.frame(elo.A = c(1500, 1500), elo.B = c(1500, 1600),
+                  wins.A = c(1, 0), k = 20)
+form <- wins.A ~ elo.A + elo.B + k(k)
+elo.prob(form, data = dat)
+elo.update(form, data = dat)
+elo.calc(form, data = dat)
+
+## ------------------------------------------------------------------------
+elo.prob(~ elo.A + elo.B, data = dat)
+
+## ------------------------------------------------------------------------
+elo.calc(wins.A ~ adjust(elo.A, 10) + elo.B + k(k), data = dat)